requirements.txt
fastapi
uvicorn[standard]
google-generativeai
crewai
requests
python-dotenv
